app bar [nfc]: Move the individual ChatNavBar buttons, too
Like the functions `ExtraButton` and `InfoButton` that wrap them,
these are only used in this one place.
